All changes are completed

+++ hashing_assignment/header_files/declarations.h
@@ -6,9 +6,9 @@ Samples getSamplesFileFromUser();
 File getFileFromUser();
 void* createArrayFromHeap(int TYPE, int size);
 char* releaseUnusedSpace(char* str);
-unsigned long int calculateFileKey(FILE* fp);
+unsigned long long int calculateFileKey(FILE* fp);
 unsigned long long int stringToKey(char* str);
-int ht_put(FILE* fp1, unsigned long int key, Entry* entry);
+int ht_put(FILE* fp1, unsigned long long int key, Entry* entry);
 void ht_create(int size);
 void ht_print();
 int isPrime(int num);
@@ -23,3 +23,4 @@ int addFileNameToSamples(char* samplesPath, char* fileName);
 int copyFileToDirectory(char* rootDirectory, File file);
 void clearScreen();
 int compareFiles(FILE* fp1, FILE* fp2);
+
